requirements.txt
google-genai
playwright
fastapi
uvicorn
python-dotenv
pydantic
pillow
